backend/requirements.txt
spotipy
python-dotenv
google-generativeai
requests
